# openapi/4-in-a-row.yaml
openapi: '3.0.3'
info:
    contact:
        name: Water
    version: '1.0.0'
    title: 4-in-a-row
    description: API zur Verwaltung von 4-Gewinnt
servers:
    - url: localhost:8000
      description: Development

tags:
    - name: Game
      description: Endpunkt zur Verwaltung von Spielen
    - name: Player
      description: Endpunkt zur Verwaltung von Spielern

paths:
    '/api/v1/games':
        get:
            summary: Lädt alle oder eine Gruppe von Spielen
            operationId: getAllGames
            x-webclient-blocking: true
            tags:
                - Game
            parameters:
                - in: query
                  name: gameStatus
                  schema:
                      $ref: '#/components/schemas/GameStatus'
                  description: Status des Spiels
            responses:
                '200':
                    description: Ok - Alle Spiele abgerufen
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Game'
                '400':
                    $ref: '#/components/responses/BadRequest'
                default:
                    $ref: '#/components/responses/InternalServerError'
        post:
            summary: Erstellt ein neues Spiel
            operationId: createGame
            x-webclient-blocking: true
            tags:
                - Game
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GameCreateRequest'
            responses:
                '201':
                    description: Created - Spiel erstellt
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Game'
                '400':
                    $ref: '#/components/responses/BadRequest'
                default:
                    $ref: '#/components/responses/InternalServerError'

    '/api/v1/games/{gameId}':
        parameters:
            - in: path
              name: gameId
              required: true
              schema:
                  type: integer
                  format: int64
              description: Einzigartige Id des Spiels
        get:
            summary: Lädt ein Spiel anhand seiner Id
            operationId: getGameById
            x-webclient-blocking: true
            tags:
                - Game
            responses:
                '200':
                    description: Ok - Spiel abgerufen
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Game'
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'
        delete:
            summary: Löscht ein Spiel anhand seiner Id
            operationId: deleteGame
            x-webclient-blocking: true
            tags:
                - Game
            responses:
                '204':
                    description: NoContent - Spiel gelöscht
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'

    '/api/v1/games/{gameId}/moves':
        parameters:
            - in: path
              name: gameId
              required: true
              schema:
                  type: integer
                  format: int64
              description: Einzigartige Id des Spiels
        post:
            summary: Ausführen eines Zuges
            operationId: makeMove
            x-webclient-blocking: true
            tags:
                - Game
            requestBody:
                description: Details zum Spielzug
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MoveRequest'
            responses:
                '200':
                    description: Ok - Spielzug erfolgreich
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Game'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'

    '/api/v1/games/{gameId}/surrender':
        parameters:
            - in: path
              name: gameId
              required: true
              schema:
                  type: integer
                  format: int64
              description: Einzigartige Id des Spiels
        post:
            summary: Spieler gibt das Spiel auf
            operationId: surrenderGame
            x-webclient-blocking: true
            tags:
                - Game
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - playerId
                            properties:
                                playerId:
                                    description: Id des aufgebenden Spielers
                                    $ref: '#/components/schemas/EntityId'
            responses:
                '204':
                    description: NoContent – Spiel erfolgreich beendet
                '400':
                    $ref: '#/components/responses/BadRequest'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'

    '/api/v1/player':
        get:
            summary: Fetched alle Spieler
            operationId: getAllPlayers
            x-webclient-blocking: true
            tags:
                - Player
            responses:
                '200':
                    description: Ok - Alle Spieler abgerufen
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Player'
                default:
                    $ref: '#/components/responses/InternalServerError'
        post:
            summary: Erstellt einen Spieler
            operationId: createPlayer
            x-webclient-blocking: true
            tags:
                - Player
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePlayerRequest'
            responses:
                '201':
                    description: Created - Spieler erstellt
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Player'
                '400':
                    $ref: '#/components/responses/BadRequest'
                default:
                    $ref: '#/components/responses/InternalServerError'

    '/api/v1/player/{playerId}':
        parameters:
            - in: path
              name: playerId
              required: true
              schema:
                  type: integer
                  format: int64
              description: Einzigartige Id des Spielers
        get:
            summary: Erhalte einen Spieler
            operationId: getPlayerById
            x-webclient-blocking: true
            tags:
                - Player
            responses:
                '200':
                    description: Ok - Spieler abgerufen
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Player'
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'
        put:
            summary: Ersetze einen Spieler
            operationId: updatePlayer
            x-webclient-blocking: true
            tags:
                - Player
            requestBody:
                description: aktualisierter Spieler
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Player'
            responses:
                '200':
                    description: Ok - Spieler aktualisiert
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Player'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'
        delete:
            summary: Löscht einen Spieler anhand seiner Id
            operationId: deletePlayer
            x-webclient-blocking: true
            tags:
                - Player
            responses:
                '204':
                    description: NoContent - Spieler gelöscht
                '404':
                    $ref: '#/components/responses/NotFound'
                default:
                    $ref: '#/components/responses/InternalServerError'

#  "/api/v1/auth/register":
#    post:
#      summary: Registriert den Nutzer mit Email und Passwort
#      operationId: register
#      x-webclient-blocking: true
#      tags:
#        - Auth
#      security:
#        - browserIdHeader: []
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              $ref: "#/components/schemas/RegisterRequest"
#      responses:
#        "201":
#          description: Created
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/RegisterResponse"
#        "400":
#          $ref: "#/components/responses/BadRequest"
#        default:
#          $ref: "#/components/responses/InternalServerError"

#  "/api/v1/auth/login":
#    post:
#      summary: Anmeldung mit Email und Passwort
#      operationId: loginMitEmailUndPasswort
#      x-webclient-blocking: true
#      tags:
#        - Auth
#      security:
#        - browserIdHeader: []
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              $ref: "#/components/schemas/LoginRequest"
#      responses:
#        "200":
#          description: Ok
#          headers:
#            Set-Cookie:
#              description: |
#                HttpOnly Cookie mit ApiKey, gültig z.B. 90 Tage
#              schema:
#                type: string
#                example: "X-Api-Key=abc123; HttpOnly; Path=/; Max-Age=7776000"
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/LoginResponse"
#        "401":
#          $ref: "#/components/responses/Unauthorized"
#        default:
#          $ref: "#/components/responses/InternalServerError"

#  "/api/v1/auth/login/google": # TODO: überarbeiten
#    post:
#      summary: Anmeldung via. Google ID Token
#      operationId: loginMitGoogle
#      x-webclient-blocking: true
#      tags:
#        - Auth
#      security:
#        - browserIdHeader: []
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              type: object
#              required:
#                - googleToken
#              properties:
#                googleToken:
#                  type: string
#                  description: Google-Id Token
#      responses:
#        "204":
#          description: NoContent
#          headers:
#            Set-Cookie:
#              description: |
#                HttpOnly Cookie mit ApiKey, gültig z.B. 90 Tage
#              schema:
#                type: string
#                example: "X-Api-Key=abc123; HttpOnly; Path=/; Max-Age=7776000"
#        "401":
#          $ref: "#/components/responses/Unauthorized"
#        default:
#          $ref: "#/components/responses/InternalServerError"

#  "/api/v1/auth/login/check":
#    get:
#      summary: Prüft, ob ein Benutzer eingeloggt ist
#      operationId: checkIsLoggedIn
#      x-webclient-blocking: true
#      tags:
#       - Auth
#      responses:
#        "200":
#          description: Ok
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/LoginResponse" # TODO: ggf. ne andere Response
#        "401":
#          $ref: "#/components/responses/Unauthorized"

#  "/api/v1/auth/logout":
#    post:
#      summary: Meldet den Nutzer ab
#      operationId: logout
#      x-webclient-blocking: true
#      tags:
#        - Auth
#      responses:
#        "204":
#          description: NoContent
#          headers:
#            Set-Cookie:
#              description: Löschen des ApiKey Cookie
#              schema:
#                type: string
#        "401":
#          $ref: "#/components/responses/Unauthorized"
#        default:
#          $ref: "#/components/responses/InternalServerError"

#  "/api/v1/debug/set_cookies": # TODO: NUR ZU DEBUG ZWECKEN
#    post:
#      summary: Setze Cookies im Browser manuell
#      operationId: setCookies
#      tags:
#        - Debug
#      x-webclient-blocking: true
#      requestBody:
#        content:
#          application/json:
#            schema:
#              properties:
#                X-Api-Key:
#                  type: string
#                  description: Wert des Cookies (der eigentliche Cookie)
#      responses:
#        200:
#          description: OK
#          headers:
#            Set-Cookie:
#              description: Cookie, das im Browser gesetzt wird
#              schema:
#                type: string
#                example: X-Api-Key=abcdef123; Path=/; HttpOnly; Secure; SameSite=None
#          content:
#            application/json:
#              schema:
#                properties:
#                  message:
#                    type: string
#                    description: Nachricht
#                  cookies:
#                    type: array
#                    description: alle gesetzen Cookies
#                    items:
#                      properties:
#                        cookie:
#                          type: string
#                          description: gesetzer Cookie
#        default:
#          $ref: "#/components/responses/InternalServerError"

components:
    responses:
        BadRequest:
            description: BadRequest
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiError'
                    example:
                        message: 'Die Anfrage konnte nicht verarbeitet werden'
                        code: 400
                        details: 'Kein QueryParameter angegeben'

        Unauthorized:
            description: Unauthorized
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiError'
                    example:
                        message: 'Authentifizierung erforderlich'
                        code: 401
                        details: 'Key fehlt oder ist ungültig'

        Forbidden:
            description: Forbidden
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiError'
                    example:
                        message: 'Zugriff verweigert'
                        code: 403
                        details: 'Benutzer hat keine Berechtigung für diese Aktion'

        NotFound:
            description: NotFound
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiError'
                    example:
                        message: 'Die angefragte Ressource wurde nicht gefunden'
                        code: 404
                        details: 'Mitglied mit ID 123 existiert nicht'

        InternalServerError:
            description: InternalServerError
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiError'
                    example:
                        message: 'Ein unerwarteter Fehler ist aufgetreten'
                        code: 500
                        details: 'NullPointerException im MemberService'

    schemas: # IDEE: Profil, mit Profilsettings, wobei die Settings/Preäferenzen als JSON direkt gespeichert werden, weil das Backend eh nicht damit arbeitet, demnach also auch nicht wissen muss was da drin steht
        EntityId:
            type: integer
            format: int64
            description: Einzigartige Id der Entität
            example: 1

        UUID:
            type: string
            format: uuid
            description: Einzigartige UUID der Entität
            example: '1bf93699-ee11-4a4a-a4a4-be7606b2de4d'

        GameCreateRequest:
            type: object
            description: Request zum Erstellen eines Spiels
            required:
                - player1Id
                - player2Id
            properties:
                player1Id:
                    description: Id des ersten Spielers
                    $ref: '#/components/schemas/EntityId'
                player2Id:
                    description: Id des zweiten Spielers
                    $ref: '#/components/schemas/EntityId'

        Game:
            type: object
            description: Response eines Spiels
            required:
                - id
                - player1
                - player2
                - status
                - currentPlayerId
                - board
            properties:
                id:
                    description: Id des Spiels
                    $ref: '#/components/schemas/EntityId'
                player1:
                    description: Spieler 1
                    $ref: '#/components/schemas/Player'
                player2:
                    description: Spieler 2
                    $ref: '#/components/schemas/Player'
                currentPlayerId:
                    description: Id des Spielers, welcher Spieler am Zug ist
                    $ref: '#/components/schemas/EntityId'
                status:
                    description: Status des Spiels
                    $ref: '#/components/schemas/GameStatus'
                result:
                    description: Ergebnis, wie das Spiel ausgegangen ist
                    $ref: '#/components/schemas/GameResult'
                board:
                    description: Spielfeld
                    $ref: '#/components/schemas/Board'

        Board:
            type: array
            items:
                type: array
                items:
                    type: integer
                    format: int32

        GameStatus:
            type: string
            description: Status des Spiels
            enum:
                - IN_PROGRESS
                - COMPLETED
                - PAUSED

        GameResult:
            type: string
            description: Ergebnis, wie das Spiel ausgegangen ist
            enum:
                - PLAYER_1_WON
                - PLAYER_2_WON
                - DRAW

        Player:
            type: object
            description: Response eines Spieler
            required:
                - id
                - username
                - email
                - registeredOn
            properties:
                id:
                    description: Id des Spieler-Profils
                    $ref: '#/components/schemas/EntityId'
                username:
                    description: Benutzername des Spielers
                    $ref: '#/components/schemas/Username'
                registeredOn:
                    description: Datum der Registrierung
                    type: string
                    format: date
                totalGames:
                    type: integer
                    format: int32
                    description: Anzahl gespielter Spiele
                    minimum: 0
                gamesWon:
                    type: integer
                    format: int32
                    description: Anzahl gewonnener Spiele
                    minimum: 0
                gamesLost:
                    type: integer
                    format: int32
                    description: Anzahl verlorener Spiele
                    minimum: 0
                gamesSurrendered:
                    type: integer
                    format: int32
                    description: Anzahl aufgegebener Spiele
                    minimum: 0
                lastPlayedOn:
                    type: string
                    format: date
                    description: Datum des letzten Spiels

        CreatePlayerRequest:
            type: object
            description: Objekt zum Erstellen eines Spielers
            required:
                - username
                - email
            properties:
                username:
                    description: Benutzername des Spielers
                    $ref: '#/components/schemas/Username'
                email:
                    description: Email des Spielers
                    $ref: '#/components/schemas/Email'

        MoveRequest:
            type: object
            required:
                - playerId
                - column
            properties:
                playerId:
                    $ref: '#/components/schemas/EntityId'
                    description: Spieler-Id des Spielers, der den Zug ausführt
                column:
                    type: integer
                    format: int32
                    minimum: 0
                    maximum: 6 # später bei variablen Spielfeldgrößen problematisch
                    description: Spalte in die der Spielstein fallen gelassen wird

        BrowserId:
            type: string
            format: uuid
            description: Eindeutige UUID des Browsers / Geräts
            example: '1bf93699-ee11-4a4a-a4a4-be7606b2de4d'

        Email:
            type: string
            format: email
            description: E-Mail Adresse
            example: 'max.musterman@gmail.com'
        Passwort:
            type: string
            format: password
            description: Passwort
            #pattern: "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[?!%&§@€$'*+<>\\.,_-])[A-Za-z\\d?!%&§@€$'*+<>\\.,_-]{8,}$"

        RegisterRequest:
            type: object
            required:
                - benutzername
                - email
                - passwort
            properties:
                benutzername:
                    description: Benutzername eines neuen Nutzers
                    $ref: '#/components/schemas/Username'
                email:
                    description: E-Mail eines neuen Nutzers
                    $ref: '#/components/schemas/Email'
                passwort:
                    description: Passwort eines neuen Nutzers
                    $ref: '#/components/schemas/Passwort'
        RegisterResponse:
            type: object
            required:
                - benutzername
                - email
            properties:
                benutzername:
                    description: Benutzername des neuen Nutzers
                    $ref: '#/components/schemas/Username'
                email:
                    description: E-Mail des neuen Nutzers
                    $ref: '#/components/schemas/Email'

        LoginRequest:
            type: object
            required:
                - benutzernameOderEmail
                - passwort
            properties:
                benutzernameOderEmail:
                    description: Benutzername oder E-Mail des Nutzers
                    anyOf:
                        - $ref: '#/components/schemas/Username'
                        - $ref: '#/components/schemas/Email'
                passwort:
                    description: Passwort des Nutzers
                    $ref: '#/components/schemas/Passwort'
        LoginResponse:
            type: object
            required:
                - benutzername
                - rollen
                - rechte
            properties:
                benutzername:
                    description: Benutzername des Nutzers
                    $ref: '#/components/schemas/Username'
                pictureLink:
                    description: Link zu einem Bild als Profilbild des Nutzers
                    $ref: '#/components/schemas/PictureLink'

        PictureLink:
            description: Link zu einem Bild
            type: string
            format: uri

        Token:
            type: object
            properties:
                token:
                    type: string
                    #pattern: "^[a-zA-Z0-9]{32}$"
                #accountId:
                #  $ref: "#/components/schemas/AccountId"

        BenutzerId:
            type: integer
            format: int32
            description: Einzigartige Id des Benutzers
            example: 1
        Username:
            type: string
            description: Benutzername
            #pattern: "^(?=.{3,30}$)(?![_.-])(?!.*[_.-]{2})[a-zA-Z0-9._-]+(?<![_.-])$"
            example: Max-Mustermann.Der_Erste12

        Steam64:
            type: integer
            description: Einzigartige Steam64-Id
            example: 76561199023920905
            minimum: 7656000000000
            maximum: 7656999999999

        DiscordBenutzerId: # TODO: ggf. als Ersatz für Discord benutzername, zumindest Rrgänzung
            type: integer
            description: Einzigartige Discord-BenutzerId

        DiscordBenutzerName:
            type: string
            #pattern: "^(?!.*\\.\\.)(?![._])[a-z0-9._]{2,32}(?<![._])$"
            example: 'paul.mcwater'
            description: Discord-Benutzername

        ApiError:
            type: object
            required:
                - statusCode
                - message
                - timestamp
                - path
            properties:
                message:
                    type: string
                    example: Statistik mit ID 42 nicht gefunden
                statusCode:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: NOT_FOUND
                timestamp:
                    type: string
                    format: date-time
                    example: '2026-02-06T17:40:00Z'
                path:
                    type: string
                    example: /api/statistiken/42
                details:
                    type: object
                    additionalProperties: true # Erlaubt die Map<String, Object>
                    description: 'Zusätzliche Fehlerdetails wie Validierungsmeldungen'

    securitySchemes:
        apiKeyCookie:
            type: apiKey
            in: cookie
            name: X-Api-Key
            description: Langlebiger API Key als HttpOnly-Cookie (90 Tage).
        browserIdHeader:
            type: apiKey
            in: header
            name: X-Browser-Id
            description: UUID zur Geräteidentifikation, muss bei jeder Anfrage mitgesendet werden
#security:
#  - apiKeyCookie: []
#    browserIdHeader: []